STRUCTURE- refactor of SmsService to be standalone, addition of TwilioController to call on SmsService and future Holiday api call for dates, and a holiday poro to make holiday objects to display in user notifications. All testing still passes. Skedush.
