+++ app/facades/holiday_facade.rb
@@ -1,0 +1,6 @@
+class HolidayFacade 
+  def self.all_holidays  
+    # (API call)Service.get_holidays
+    # map the call and make an array of holiday poros
+  end
+end
