+++ app/poros/holiday.rb
@@ -1,0 +1,11 @@
+class Holiday 
+  # attr_reader :id,
+  #             :title,
+  #             :date
+
+  # def initialize(data)
+  #   # @id = data[:id]
+  #   # @title = data[:title]
+  #   # @date = data[:date]
+  # end
+end
